+++ internal/api/web/_app/immutable/entry/start.Ddndua8c.js
@@ -1,0 +1,1 @@
+import{o as e,t}from"../chunks/BFsL76vj.js";export{e as load_css,t as start};